rpi-client/requirements.txt
websockets>=12.0
requests>=2.31.0
RPi.GPIO>=0.7.1; platform_machine == "armv7l" or platform_machine == "aarch64"
gpiozero>=2.0; platform_machine == "armv7l" or platform_machine == "aarch64"
schedule>=1.2.0
python-dotenv>=1.0.0

